Find the entire SDK instead.
