+++ msbuild/Xamarin.Localization.MSBuild/MSBStrings.resx
@@ -1542,9 +1542,6 @@
     </data>
 
     <data name="E7120" xml:space="preserve">
-        <value>The mlaunch tool doesn't exist: '{0}'</value>
-        <comment>
-            {0}: the path to the mlaunch tool
-        </comment>
+        <value>The SDK directory {0} doesn't exist.</value>
     </data>
 </root>
