+++ msbuild/Xamarin.iOS.Tasks.Windows/Xamarin.iOS.Common.After.targets
@@ -13,7 +13,7 @@ Copyright (C) 2011-2013 Xamarin. All rights reserved.
 -->
 <Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">	
 	<UsingTask TaskName="Xamarin.MacDev.Tasks.FindILLink" AssemblyFile="$(CoreiOSSdkDirectory)Xamarin.iOS.Tasks.dll" />
-	<UsingTask TaskName="Xamarin.MacDev.Tasks.FindMlaunch" AssemblyFile="$(CoreiOSSdkDirectory)Xamarin.iOS.Tasks.dll" />
+	<UsingTask TaskName="Xamarin.MacDev.Tasks.LocateXamarinSdk" AssemblyFile="$(CoreiOSSdkDirectory)Xamarin.iOS.Tasks.dll" />
 	<UsingTask TaskName="Xamarin.MacDev.Tasks.Unzip" AssemblyFile="$(CoreiOSSdkDirectory)Xamarin.iOS.Tasks.dll" />
 	<UsingTask TaskName="Xamarin.MacDev.Tasks.WriteItemsToFile" AssemblyFile="$(CoreiOSSdkDirectory)Xamarin.iOS.Tasks.dll" />
 	<UsingTask TaskName="Xamarin.MacDev.Tasks.Zip" AssemblyFile="$(CoreiOSSdkDirectory)Xamarin.iOS.Tasks.dll" />
@@ -340,16 +340,16 @@ Copyright (C) 2011-2013 Xamarin. All rights reserved.
 		</ItemGroup>
 	</Target>
 
-	<Target Name="_FindMlaunchPath">
-		<FindMlaunch
+	<Target Name="_LocateXamarinSdk">
+		<LocateXamarinSdk
 			SessionId="$(BuildSessionId)"
-			Condition="'$(IsMacEnabled)' == 'true' And '$(_RemoteMLaunchPath)' == ''"
-			KeepTemporaryOutput="$(FindMlaunchPathKeepKeepTemporaryOutput)"
+			Condition="'$(IsMacEnabled)' == 'true' And '$(_RemoteXamarinSdkRootDirectory)' == ''"
+			KeepTemporaryOutput="$(LocateXamarinSdkKeepKeepTemporaryOutput)"
 			RuntimeIdentifier="$(RuntimeIdentifier)"
 			TargetFrameworkMoniker="$(_ComputedTargetFrameworkMoniker)"
 			>
-			<Output TaskParameter="MlaunchPath" PropertyName="_RemoteMLaunchPath" />
-		</FindMlaunch>
+			<Output TaskParameter="XamarinSdkRootDirectory" PropertyName="_RemoteXamarinSdkRootDirectory" />
+		</LocateXamarinSdk>
 	</Target>
 
 	<!-- Overrides Core SDK target to remote the ILLink execution -->
